🐛 Fix:#80 - 코멘트 수정사항 반영
