+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18.3.1",
     "react-datepicker": "^7.5.0",
     "react-dom": "^18.3.1",
+    "react-intersection-observer": "^9.13.1",
     "react-quill": "^2.0.0",
     "react-quill-new": "^3.3.3",
     "zustand": "^5.0.1"
