+++ src/api/workService.ts
@@ -1,4 +1,4 @@
-import { deleteRequest, getRequest } from './api';
+import { deleteRequest, getRequest, postRequest } from './api';
 
 export const getWorkDetail = async (workId: string) => {
   try {
@@ -14,5 +14,23 @@ export const deleteWorkDetail = async (workId: string) => {
     await deleteRequest(`/works/${workId}`);
   } catch (error) {
     throw new Error('Failed to delete works.');
+  }
+};
+
+export const getFeedbackList = async (workId: string, page: number, pageSize: number = 4) => {
+  try {
+    const res = await getRequest(`/works/${workId}/feedbacks?page=${page}&pageSize=${pageSize}`);
+    return res.data;
+  } catch (error) {
+    throw new Error('Failed to get feedback');
+  }
+};
+
+export const postFeedback = async (workId: string, content: string) => {
+  try {
+    const res = await postRequest(`/works/${workId}/feedbacks`, { content });
+    return res.data;
+  } catch (error) {
+    throw new Error('Failed to create feedback');
   }
 };
